fixing collective saving and search functions. Added list searching and search by name and id
